# backend/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppGetHealthController_getHealth","summary":"Return the application's health","parameters":[],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppGetHealthResponseDTO"}}},"description":""}}}},"/user/create":{"post":{"operationId":"UserCreateUserController_create","summary":"Create a new user from his name","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateUserBodyDTO"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateUserResponseDTO"}}},"description":""},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateUserResponseDTO"}}}}}}},"/messenger/create":{"post":{"operationId":"MessengerCreateConversationController_create","summary":"Create a new conversation","parameters":[],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessengerCreateConversationResponseDTO"}}},"description":""},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessengerCreateConversationResponseDTO"}}}}}}},"/messenger/{messengerId}/send-message":{"post":{"operationId":"MessengerSendMessageToAssistantController_sendMessage","summary":"Create a new conversation","parameters":[{"name":"messengerId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/MessengerSendMessageToAssistantBodyDTO"}}}},"responses":{"200":{"content":{"text/event-stream":{"schema":{"type":"string"}}},"description":""},"201":{"description":""}}}},"/messenger/{messengerId}/messages":{"get":{"operationId":"MessengerGetMessagesController_getMessages","summary":"Create a new conversation","parameters":[{"name":"messengerId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessengerGetMessagesResponseDTO"}}},"description":""}}}}},"info":{"title":"Swagger API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AppGetHealthResponseDTO":{"type":"object","properties":{"status":{"type":"string"}},"required":["status"]},"UserCreateUserBodyDTO":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UserCreateUserResponseDTO":{"type":"object","properties":{"sessionId":{"type":"string"}},"required":["sessionId"]},"MessengerCreateConversationResponseDTO":{"type":"object","properties":{"messengerId":{"type":"string"}},"required":["messengerId"]},"MessengerSendMessageToAssistantBodyDTO":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"MessageDTO":{"type":"object","properties":{"message":{"type":"string"},"sender":{"type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["message","sender","createdAt"]},"MessengerGetMessagesResponseDTO":{"type":"object","properties":{"messages":{"type":"array","items":{"$ref":"#/components/schemas/MessageDTO"}}},"required":["messages"]}}}}